API_Recommendation/requirements.txt
flask
pandas
numpy
pymongo
python-dotenv
dnspython